requirements.txt
kafka-python
fastapi
psycopg2-binary
uvicorn